+++ app/Controllers/HomeController.php
@@ -1,14 +1,9 @@
 <?php
 
-declare(strict_types=1);
-
 namespace App\Controllers;
 
-use Codemonster\Annabel\Http\Response;
+use Codemonster\Http\Response;
 
-/**
- * The default controller that renders the home page.
- */
 class HomeController
 {
     public function index(): Response