+++ bootstrap/app.php
@@ -1,12 +1,6 @@
 <?php
 
-declare(strict_types=1);
-
 use Codemonster\Annabel\Application;
-
-/**
- * Bootstrap the Annabel application and load routes.
- */
 
 $baseDir = __DIR__ . '/..';